refactor: App 컴포넌트 Header + Wordle로 재구성
