+++ README.md
@@ -49,7 +49,7 @@ https://wordle-challenge-psi.vercel.app/
 [x] 엔터 눌렀을때 애니메이션   
 [x] 글자 입력되었을때 애니메이션 동작   
 [x]  정답일때와 아닐때 애니메이션 동작   
-[ ] 모달 css 처리   
+[x] 모달 css 처리   
 [ ] 다크/안다크 모드 추가   
 [ ] 영어/한국어 버전 추가   
 [ ] 반응형 작업
