+++ src/components/GameBoard.tsx
@@ -14,7 +14,6 @@ const GameBoard = ({
   currentInput,
   guessEvaulations,
 }: Props) => {
-  console.log(guesses);
   return (
     <GameBoardWrapper>
       {guesses.map((guess, wordIndex) => {
