+++ src/components/Header.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Header() {
+  return <div></div>;
+}
