+++ src/components/index.ts
@@ -2,8 +2,12 @@ import KeyBoard from "./KeyBoard";
 import GameBoard from "./GameBoard";
 import NavFadeModal, { NavModalPortal } from "./NavFadeModal";
 import ResultModal, { ResultModalPortal } from "./ResultModal";
+import Header from "./Header";
+import Wordle from "./Wordle";
 
 export {
+  Wordle,
+  Header,
   KeyBoard,
   GameBoard,
   NavFadeModal,
